proyectos_python/file_upload/requirements.txt
snowflake-connector-python[pandas]==3.0.2
openpyxl==3.1.2
pyinstaller==5.10.1
Unidecode==1.3.6